feat(browser-stopped): PipeCompare → Polling + LongPoll + WebSocketStream
§3 decomposition. Each protocol now lives in its own WidgetShell with a
single play button — one verb per shell, one teaching claim per widget.
Bridge prose between the three carries the comparison the old combined
canvas used to make visually. The closing-thesis HL on "refusing to
finish the question" lands after the WebSocket widget, so the reader
has watched all three protocols before the unifying claim arrives.

Shared helpers _pipe-sim.ts (pure simulators) and _pipe-canvas.tsx
(canvas + rAF clock hook) carry the verbatim simulation contracts so
the new widgets are visually identical to the rows of PipeCompare. The
clock hook cancels its rAF callback the moment a widget pauses; idle
widgets register zero callbacks.
