+++ app/posts/the-browser-stopped-asking/widgets/index.ts
@@ -1,4 +1,6 @@
-export { PipeCompare } from "./PipeCompare";
+export { Polling } from "./Polling";
+export { LongPoll } from "./LongPoll";
+export { WebSocketStream } from "./WebSocketStream";
 export { UpgradeHandshake } from "./UpgradeHandshake";
 export { ReconnectGap } from "./ReconnectGap";
 export { CostMatrix } from "./CostMatrix";
